Moved command options to their own files
